LayoutTests/imported/w3c/web-platform-tests/css/css-borders/corner-shape/corner-top-valid-expected.txt


FAIL e.style['corner-top'] = "10px bevel" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top'] = "round 10% / 2rem squircle" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top'] = "normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top'] = "4px 2% round / bevel 1em" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['corner-top'] = "10px bevel" should set border-top-left-radius assert_equals: border-top-left-radius should be canonical expected "10px" but got ""
FAIL e.style['corner-top'] = "10px bevel" should set border-top-right-radius assert_equals: border-top-right-radius should be canonical expected "10px" but got ""
FAIL e.style['corner-top'] = "10px bevel" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-top'] = "10px bevel" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-top'] = "10px bevel" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['corner-top'] = "4px 2% round / 1em bevel" should set border-top-left-radius assert_equals: border-top-left-radius should be canonical expected "4px 2%" but got ""
FAIL e.style['corner-top'] = "4px 2% round / 1em bevel" should set border-top-right-radius assert_equals: border-top-right-radius should be canonical expected "1em" but got ""
FAIL e.style['corner-top'] = "4px 2% round / 1em bevel" should set corner-top-left-shape assert_equals: corner-top-left-shape should be canonical expected "round" but got ""
FAIL e.style['corner-top'] = "4px 2% round / 1em bevel" should set corner-top-right-shape assert_equals: corner-top-right-shape should be canonical expected "bevel" but got ""
FAIL e.style['corner-top'] = "4px 2% round / 1em bevel" should not set unrelated longhands assert_true: expected true got false
